Page Splitted to Home, State Dashboaed and Footer
