+++ src/Pages/Footer.jsx
@@ -1,0 +1,15 @@
+import { styles } from "../Styles";
+
+function Footer() {
+  return (
+    <div>
+      {/* Footer */}
+      <footer style={styles.footer}>
+        <p>Data Powered by Information Kerala Mission</p>
+        <p>Contact: mail@ikm.gov.in</p>
+      </footer>
+    </div>
+  );
+}
+
+export default Footer;
